[SPARK-42115][SQL] Push down limit through Python UDFs
### What changes were proposed in this pull request?

This PR adds cases in LimitPushDown to push limits through Python UDFs. In order to allow for this, we need to call LimitPushDown in SparkOptimizer after the "Extract Python UDFs" batch. We also add PushProjectionThroughLimit afterwards in order to plan CollectLimit.

### Why are the changes needed?

Right now, LimitPushdown does not push limits through Python UDFs, which means that expensive Python UDFs can be run on potentially large amounts of input. This PR adds this capability, while ensuring that a GlobalLimit - LocalLimit pattern stays at the top in order to trigger the CollectLimit code path.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Added a UT.

Closes #39842 from kelvinjian-db/SPARK-42115-limit-through-python-udfs.

Authored-by: Kelvin Jiang <kelvin.jiang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>