[MINOR][DOCS][PYTHON][PS] Fix the `.groupby()` method docstring
### What changes were proposed in this pull request?

Update the docstring for the `.groupby()` method.

### Why are the changes needed?

The `.groupby()` method accept a list of columns (or a single column), and a column is defined by a `Series` or name (`Label`). It's a bit confusing to say "using a Series of columns", because `Series` (capitalized) is a specific object that isn't actually used/reasonable to use here.

### Does this PR introduce _any_ user-facing change?

Yes (documentation)

### How was this patch tested?

N/A

Closes #38625 from deepyaman/patch-3.

Authored-by: Deepyaman Datta <deepyaman.datta@utexas.edu>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>