[SPARK-41600][SPARK-41623][SPARK-41612][CONNECT] Implement Catalog.cacheTable, isCached and uncache
### What changes were proposed in this pull request?

This PR adds three API below to Spark Connect
- `Catalog.isCached`
- `Catalog.cacheTable`
- `Catalog uncacheTable`

### Why are the changes needed?

These were not added because of the design concern (in its behaviour). However, we should provide the same API compatibility and behaivours with the regular PySpark in any event. So these are proposed back.

### Does this PR introduce _any_ user-facing change?

No to end users.
Yes to the dev because it adds three new API in Spark Connect.

### How was this patch tested?

Unittests were added.

Closes #39919 from HyukjinKwon/SPARK-41600-SPARK-41623-SPARK-41612.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>