[SPARK-41716][CONNECT] Rename _catalog_to_pandas to _execute_and_fetch in Catalog
### What changes were proposed in this pull request?

This PR proposes to rename `_catalog_to_pandas` to `_execute_and_fetch`, and remove the JIRA in the comment.

### Why are the changes needed?

I tried to factor this out to the client but I realised that there won't be too much difference.
In addition, we need a new design for Catalog API so I proposed to remove the JIRA for now.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing test cases should cover.

Closes #39920 from HyukjinKwon/SPARK-41716-1.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>