[SPARK-40819][SQL][FOLLOWUP] Update SqlConf version for nanosAsLong configuration
As requested by HyukjinKwon in https://github.com/apache/spark/pull/38312  NB: This change needs to be backported

### What changes were proposed in this pull request?

Update version set for "spark.sql.legacy.parquet.nanosAsLong"  configuration in SqlConf. This update is required because the previous PR set version to `3.2.3` which has already been released. Updating to version `3.2.4` will correctly reflect when this configuration element was added

### Why are the changes needed?

Correctness and to complete SPARK-40819

### Does this PR introduce _any_ user-facing change?

No, this is merely so this configuration element has the correct version

### How was this patch tested?

N/A

Closes #39943 from awdavidson/SPARK-40819_sql-conf.

Authored-by: awdavidson <54780428+awdavidson@users.noreply.github.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>