[SPARK-42338][CONNECT] Add details to non-fatal errors to raise a proper exception in the Python client
### What changes were proposed in this pull request?

Adds details to non-fatal errors to raise a proper exception in the Python client, which makes `df.sample` raise `IllegalArgumentException` as same as PySpark, except for the timing that is delayed to call actions.

### Why are the changes needed?

Currently `SparkConnectService` does not add details for `NonFatal` exceptions to the `PRCStatus`, so the Python client can't detect the exception properly and raises `SparkConnectGrpcException` instead.

It also should have the details for the Python client.

### Does this PR introduce _any_ user-facing change?

Users will see a proper exception when they call `df.sample` with illegal arguments, but in a different timing.

### How was this patch tested?

Enabled `DataFrameParityTests.test_sample`.

Closes #39957 from ueshin/issues/SPARK-42338/sample.

Authored-by: Takuya UESHIN <ueshin@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>