[SPARK-42418][DOCS][PYTHON] PySpark documentation updates to improve discoverability and add more guidance
### What changes were proposed in this pull request?
Updates to the PySpark documentation web pages that help users choose which API to use when and make it easier to discover relevant content and navigate the documentation pages. This is the first of a series of updates.

### Why are the changes needed?
The PySpark documentation web site does not do enough to help users choose which API to use and it is not easy to navigate.

### Does this PR introduce _any_ user-facing change?
Yes, the user facing PySpark documentation is updated.

### How was this patch tested?
Built and tested the PySpark documentation web site locally.

Closes #39992 from allanf-db/pyspark_doc_updates.

Authored-by: Allan Folting <allan.folting@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>