[SPARK-42406][PROTOBUF] Fix recursive depth setting for Protobuf functions
### What changes were proposed in this pull request?

This fixes how setting for limiting recursive depth is handled in Protobuf functions (`recursive.fields.max.depth`).
Original documentation says as setting it to '0' removes the recursive field. But we never did that. We allow at least once. E.g. schema for recursive message 'EventPerson' does not change between the settings '0' or '1'.

This fixes it by requiring the max depth to be at least 1. It also fixes how the recursion enfored.
Updated the tests and added an extra test with new protobuf 'EventPersonWrapper'.

I will annotate the diff inline pointing to main fixes.

### Why are the changes needed?

This fixes a bug with enforcing `recursive.fields.max.depth` and clarifies more in the documentation.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
- Unit tests

Closes #40011 from rangadi/recursive-depth.

Authored-by: Raghu Angadi <raghu.angadi@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>