[SPARK-42465][CONNECT] ProtoToPlanTestSuite should use analyzed plans instead of parsed plans
### What changes were proposed in this pull request?
This makes `ProtoToPlanTestSuite` use analyzed plans instead of parsed plans.

### Why are the changes needed?
This is to increase the fidelity of the `ProtoToPlanTestSuite`, especially since we are going to be adding functions. Functions are special because the spark connect planner leaves them unresolved, the actual binding only happens in the analyzer. Without running the analyzer we would not know if the bindings are correct.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
It is a test change.

Closes #40056 from hvanhovell/SPARK-42465.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>