[SPARK-42406][SQL] Fix check for missing required fields of to_protobuf
### What changes were proposed in this pull request?

Protobuf serializer (used in `to_protobuf()`) should error if non-nullable fields (i.e. protobuf `required` fields) are present in the schema of the catalyst record being converted to a protobuf.

But `isNullable()` method used for this check returns opposite (see PR comment in the diff).  As a result, Serializer incorrectly requires the fields that are optional. This PR fixes this check (see PR comment in the diff).

This also requires corresponding fix for couple of unit tests. In order use a Protobuf message with a `required` field, Protobuf version 2 file `proto2_messages.proto` is added.
Two tests are updated to verify missing required fields results in an error.

### Why are the changes needed?

This is need to fix a bug where we were incorrectly enforcing a schema check on optional fields rather than on required fields.

### Does this PR introduce _any_ user-facing change?

It fixes a bug, and gives more flexibility for user queries.

### How was this patch tested?
 - Updated unit tests

Closes #40080 from rangadi/fix-required-field-check.

Authored-by: Raghu Angadi <raghu.angadi@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>