[SPARK-42531][CONNECT] Scala Client Add Collections Functions
### What changes were proposed in this pull request?
This PR adds all the collections functions to `functions.scala` for Scala client. This is the last PR large functions PR, there are a few functions missing, these will be added later.

### Why are the changes needed?
We want the Scala client to have API parity with the existing API

### Does this PR introduce _any_ user-facing change?
Yes, it adds functions to the Spark Connect Scala Client.

### How was this patch tested?
Added tests to `PlanGenerationTestSuite` and to `ProtoToPlanTestSuite`. I have added a few tests to `ClientE2ETestSuite` for lambda functions (to make sure name scoping works) and the array shuffle function (non-deterministic, hard to test with golden files).

Closes #40130 from hvanhovell/SPARK-42531.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>