[SPARK-42543][CONNECT] Specify protocol for UDF artifact transfer in JVM/Scala client
### What changes were proposed in this pull request?

This PR introduces a new client-streaming RPC service "AddArtifacts" to handle the transfer of artifacts from the client to the server. New message types `AddArtifactsRequest` and `AddArtifactsResponse` are added that specify the format of artifact transfer.

An artifact is defined by its `name` and `data` fields.

- `name`
  - The name of the artifact is expected in the form of a "Relative Path" that is made up of a sequence of directories and the final file element.
  - Examples of "Relative Path"s: `jars/test.jar`,  `classes/xyz.class`, `abc.xyz`, `a/b/X.jar`.
  - The server is expected to maintain the hierarchy of files as defined by their name. (i.e The relative path of the file on the server's filesystem will be the same as the name of the provided artifact).
- `data`
  - The raw data of the artifact.

The intention behind the `name` format is to add extensibility to the approach. Through this scheme, the server can maintain the hierarchy/grouping of files in any way the client specifies as well as transfer different "forms" of artifacts without needing any updates to the protocol/code itself.

The protocol supports batching and chunking (due to gRPC size limits) of artifacts as required.

### Why are the changes needed?

In the decoupled client-server architecture of Spark Connect, a remote client may use a local JAR or a new class in their UDF that may not be present on the server. To handle these cases of missing "artifacts", a protocol for artifact transfer is needed to move the required artifacts from the client side over to the server side.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

N/A

Closes #40147 from vicennial/artifactProtocol.

Authored-by: vicennial <venkata.gudesa@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>