[SPARK-41725][CONNECT] Eager Execution of DF.sql()
### What changes were proposed in this pull request?

This patch allows for eager execution of SQL statements using the Spark Connect Data Frame API.  The implementation of the patch is as follows: When `spark.sql` is called, the client sends a command to the server including the SQL statement. The server will evaluate the query and execute the side-effects if necessary. If the query was a command it will return the results as a `Relaiton.LocalRelation` back to the client otherwise it will return a `Relation.SQL` to the client. The client then simply forwards the received payload to the next operator.

### Why are the changes needed?
Compatibility

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT

Closes #40160 from grundprinzip/eager_sql_v2.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>