[SPARK-42458][CONNECT][PYTHON] Fixes createDataFrame to support DDL string as schema
### What changes were proposed in this pull request?

Fixes `createDataFrame` to support DDL string as schema.

### Why are the changes needed?

Currently DDL string as schema is ignored when the data is Python objects and the inference fails:

```py
>>> spark.createDataFrame([(100, None)], "age INT, name STRING").show()
Traceback (most recent call last):
...
ValueError: Some of types cannot be determined after inferring, a StructType Schema is required in this case
```

### Does this PR introduce _any_ user-facing change?

The DDL string as schema will not be ignored when the schema can't be inferred from the given data.

```py
>>> spark.createDataFrame([(100, None)], "age INT, name STRING").show()
+---+----+
|age|name|
+---+----+
|100|null|
+---+----+
```

### How was this patch tested?

Enabled related tests.

Closes #40240 from ueshin/issues/SPARK-42458/schema_str.

Authored-by: Takuya UESHIN <ueshin@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>