[SPARK-42493][DOCS][PYTHON] Make Python the first tab for code examples - Spark SQL, DataFrames and Datasets Guide
### What changes were proposed in this pull request?
Making Python the first tab for code examples in the Spark SQL, DataFrames and Datasets Guide.

### Why are the changes needed?
Python is the easiest approachable and most popular language so this change moves it to the first tab (showing by default) for code examples.

### Does this PR introduce _any_ user-facing change?
Yes, the user facing Spark documentation is updated.

### How was this patch tested?
I built the website locally and manually tested the pages.

Closes #40087 from allanf-db/spark_docs.

Authored-by: Allan Folting <allan.folting@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>