[SPARK-42687][SS] Better error message for the unsupport `pivot` operation in Streaming
`pivot` is an unsupported operation in structured streaming but produces a bad error message that is quite misleading.

The following is the current error message for the pivot in SS:
```
AnalysisException: Queries with streaming sources must be executed with writeStream.start();
```

We should report that an unsupported operation is happening instead of stating another operation should start the stream.

### What changes were proposed in this pull request?

Throw AnalysisExeception when `pivot` is called in Structure Streaming

### Why are the changes needed?

Better error message for a better user journey

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

UT

Closes #40306 from huanliwang-db/error-msg.

Authored-by: Huanli Wang <huanli.wang@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>