[SPARK-42656][CONNECT][FOLLOWUP] Improve the script to start spark-connect server
### What changes were proposed in this pull request?
Make the server to start with Hive dependency and make the script to build with the correct scala version set in pom files.
Also allow the server to take any extra spark configurations.

### Why are the changes needed?
Adding hive dependency to allow some tests of write ops easier.
Allow the server to start with the same Scala version set in the project.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Manually tested.

Closes #40303 from zhenlineo/improve-scripts.

Authored-by: Zhen Li <zhenlineo@users.noreply.github.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>