[SPARK-42722][CONNECT][PYTHON] Python Connect def schema() should not cache the schema
### What changes were proposed in this pull request?

As of now that Connect Python client cache the schema when calling `def schema()`. However this might cause stale data issue. For example:
```
1. Create table
2. table.schema
3. drop table and recreate the table with different schema
4. table.schema // now this is incorrect
```

### Why are the changes needed?

Fix the behavior when the cached schema could be stale.

### Does this PR introduce _any_ user-facing change?

This is actually a fix that users now can always see the most up-to-dated schema.

### How was this patch tested?

Existing UT

Closes #40343 from amaliujia/disable_cache.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>