[SPARK-42689][CORE][SHUFFLE] Allow ShuffleDriverComponent to declare if shuffle data is reliably stored
### What changes were proposed in this pull request?

Currently, if there is an executor node loss, we assume the shuffle data on that node is also lost. This is not necessarily the case if there is a shuffle component managing the shuffle data and reliably maintaining it (for example, in distributed filesystem or in a disaggregated shuffle cluster).

### Why are the changes needed?

Downstream projects have patches to Apache Spark in order to workaround this issue, for example Apache Celeborn has [this](https://github.com/apache/incubator-celeborn/blob/main/assets/spark-patch/RSS_RDA_spark3.patch).

### Does this PR introduce _any_ user-facing change?

Enhances the `ShuffleDriverComponents` API, but defaults to current behavior.

### How was this patch tested?

Existing unit tests, and added more tests.

Closes #40307 from mridulm/support-hook-for-reliable-shuffle.

Authored-by: Mridul Muralidharan <mridulatgmail.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>