[SPARK-42685][CORE] Optimize Utils.bytesToString routines
### What changes were proposed in this pull request?

Optimize `Utils.bytesToString`. Arithmetic ops on `BigInt` and `BigDecimal` are order(s) of magnitude slower than the ops on primitive types. Division is an especially slow operation and it is used en masse here.

To avoid heating up the Earth while formatting byte counts for human consumption we observe that most formatting operations are not in the 10s of EiBs but on counts that fit in 64-bits and use (fastpath) 64-bit operations to format them.

### Why are the changes needed?
Use of `Utils.bytesToString` is prevalent through the codebase and they are mainly used in logs. If the logs are emitted then this becomes a heavyweight operation.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

Closes #40301 from alkis/faster-byte-to-string.

Authored-by: Alkis Evlogimenos <alkis.evlogimenos@databricks.com>
Signed-off-by: Sean Owen <srowen@gmail.com>