[SPARK-42667][CONNECT][FOLLOW-UP] SparkSession created by newSession should not share the channel
### What changes were proposed in this pull request?

SparkSession created by newSession should not share the channel. This is because that a SparkSession might be called `stop` in which the channel it uses will be shutdown. If the channel is shared, other non-stop SparkSession that is sharing this channel will get into trouble.

### Why are the changes needed?

This fixes the issue when one SparkSession is stopped to cause other active SparkSession not working in Spark Connect.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing UT

Closes #40346 from amaliujia/rw-session-do-not-share-channel.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>