[SPARK-42755][CONNECT] Factor literal value conversion out to `connect-common`
### What changes were proposed in this pull request?
Factor literal value conversion out to `connect-common`.

### Why are the changes needed?
when trying to build protos of literal array in the server side for ml, I found we have two implementations:

`LiteralExpressionProtoConverter. toConnectProtoValue`  in server module, but it doesn't support array;
`LiteralProtoConverter. toLiteralProto` in client module, it support more types;

We'd better factor it out to common module, so that both client and server can leverage it.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
existing UT

Closes #40375 from zhengruifeng/connect_mv_literal_common.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>