[SPARK-42577][CORE] Add max attempts limitation for stages to avoid potential infinite retry
### What changes were proposed in this pull request?
Currently a stage will be resubmitted in a few scenarios:
1. Task failed with `FetchFailed` will trigger stage re-submit;
2. Barrier task failed;
3. Shuffle data loss due to executor/host decommissioned;

For the first 2 scenarios, there is a config `spark.stage.maxConsecutiveAttempts` to limit the retry times. While for the 3rd scenario, there'll be potential risks for inifinite retry if there are always executors hosting the shuffle data from successful tasks got killed/lost, the stage will be re-run again and again.

To avoid the potential risk, the proposal in this PR is to add a new config `spark.stage.maxConsecutiveAttempts` to limit the overall max attempts number for each stage, the stage will be aborted once the retry times beyond the limitation.

### Why are the changes needed?
To avoid the potential risks for stage infinite retry.

### Does this PR introduce _any_ user-facing change?
Added limitation for stage retry times, so jobs may fail if they need to retry for mutiplte times beyond the limitation.

### How was this patch tested?
Added new UT.

Closes #40286 from ivoson/SPARK-42577.

Authored-by: Tengfei Huang <tengfei.h@gmail.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>