[SPARK-42799][BUILD] Update SBT build `xercesImpl` version to match with `pom.xml`
### What changes were proposed in this pull request?

This PR aims to update `XercesImpl` version to `2.12.2` from `2.12.0` in order to match with the version of `pom.xml`.

https://github.com/apache/spark/blob/88e72388b2cb1e1531ddc215693899eed79f1303/pom.xml#L1429-L1433

### Why are the changes needed?

When we updated this version via SPARK-39183, we missed to update `SparkBuild.scala`.

- https://github.com/apache/spark/pull/36544

### Does this PR introduce _any_ user-facing change?

No, this is a dev-only change because the release artifact' dependency is managed by Maven.

### How was this patch tested?

Pass the CIs.

Closes #40431 from dongjoon-hyun/SPARK-42799.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>