[SPARK-42791][SQL] Create a new golden file test framework for analysis
### What changes were proposed in this pull request?

This PR updates the `SQLQueryTestSuite` to also consume the same input SQL queries from the input files in second pass and then perform analysis and generate the string representation of the analyzed plans, in a separate new directory. It works much the same way as the previous `SQLQueryTestSuite` behavior except that it now also produces analyzed plan string instead of just query results in the output golden files.

### Why are the changes needed?

This framework will help us guard against bugs in future development by showing a clear signal when analyzer updates result in changes to the query plans for a body of test queries. PR authors and reviewers will be able to see the diffs for all changed plans during the review, and any unintentional plan changes will act as a signal for PR authors to adjust their code to prevent the changes from happening.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This test adds a new test suite and initial golden file. The new test suite passes as of this PR.

Closes #40449 from dtenedor/analyzer-golden-files.

Authored-by: Daniel Tenedorio <daniel.tenedorio@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>